inverted which input is disabled
